hocus-pocus is a framework for terminal apps.
This framework takes advantage of notty escape codes. It will have a
"polyfill" capability to emulate the features of notty terminals in
ANSI terminals when possible. This capability is not yet implemented.

This commit contains the LineBuffer structure, which is a line-oriented
interface based on the local echo capabilities of a notty terminal. It
currently does not implement any history capabilities or other advanced
server-side features, or even writing to stderr.